+++ apps/production/src/project/project.service.ts
@@ -1192,4 +1192,24 @@ export class ProjectService {
   deleteProjects(ids: string[]): void {
     this.projectsRepository.delete(ids)
   }
+
+  async deleteProjectsData(projectIds: string[]): Promise<void> {
+    const projectIdStrings = projectIds.map(id => `'${id}'`).join(', ')
+
+    const query = `
+      DELETE FROM analytics
+      WHERE pid IN (${projectIdStrings});
+
+      DELETE FROM captcha
+      WHERE pid IN (${projectIdStrings});
+
+      DELETE FROM customEV
+      WHERE pid IN (${projectIdStrings});
+
+      DELETE FROM performance
+      WHERE pid IN (${projectIdStrings});
+    `
+
+    await clickhouse.query(query).toPromise()
+  }
 }
